Add GUI elements to app.py
Integrate GUI-related methods and functionalities from `src/gui.py` into `src/app.py`.

* **Initialization and Widgets**:
  - Add `create_widgets` method to initialize GUI elements.
  - Modify `__init__` method to include root window setup and widget creation.
  - Add imports for `tkinter`, `ttk`, and other necessary modules.

* **Menu and Dark Mode**:
  - Add `create_menu` method to create the menu bar.
  - Add `toggle_dark_mode` and `apply_theme` methods to handle dark mode toggling and theme application.

* **Tabs and GUI Elements**:
  - Add methods to create various tabs: `create_logs_tab`, `create_exploits_tab`, `create_communication_tab`, `create_device_control_tab`, `create_target_scanning_tab`, and `create_ai_model_tab`.
  - Add methods for tab functionalities: `refresh_logs`, `load_exploits`, `run_exploit`, `send_message`, `deploy_exploit`, `scan_targets`, and `predict`.

* **Dialogs and Feedback**:
  - Add `show_about` and `show_help` methods to display information dialogs.
  - Add methods for feedback: `report_issue` and `suggest_improvement`.

* **Additional Functionalities**:
  - Add methods for user login, session management, and 2FA: `login`, `authenticate_user`, `get_user_role`, `check_session_timeout`, and `implement_2fa`.
  - Add methods for secure communication, DDNS setup, reverse DNS tunneling, and chatbot integration.

* **Main Execution**:
  - Add main execution block to initialize and run the application.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ProjectZeroDays/zero-click-exploits?shareId=XXXX-XXXX-XXXX-XXXX).
