--- src/gui.py
@@ -1,490 +1,0 @@
-# This file contains the main GUI implementation for the app
-
-import tkinter as tk
-from tkinter import ttk, messagebox
-import os
-import subprocess
-import re
-import shodan
-import nmap
-import logging
-import json
-import base64
-from cryptography.fernet import Fernet
-from tkinter.simpledialog import askstring
-import requests
-from project_red_sword import Chatbot
-from ai_model import AIDeploymentModel
-from tkinter import dnd
-from tkinter import tooltip
-from advanced_decryption import AdvancedDecryption
-from advanced_malware_analysis import AdvancedMalwareAnalysis
-from advanced_social_engineering import AdvancedSocialEngineering
-from adware_manager import AdwareManager
-from ai_red_teaming import AIRedTeaming
-from alerts_notifications import AlertsNotifications
-from android_exploit import AndroidExploit
-from apt_simulation import APTSimulation
-from automated_incident_response import AutomatedIncidentResponse
-from blockchain_logger import BlockchainLogger
-from botnet_manager import BotnetManager
-from data_exfiltration import DataExfiltration
-from data_visualization import DataVisualization
-from deepseek_cody_integration_manager import DeepSeekCodyIntegrationManager
-from device_fingerprinting import DeviceFingerprinting
-from dns_manager import DNSManager
-from download_manager import DownloadManager
-from exploit_payloads import ExploitPayloads
-from fuzzing_engine import FuzzingEngine
-from identity_manager import IdentityManager
-from ios_exploit import IOSExploit
-from iot_exploitation import IoTExploitation
-from linux_exploit import LinuxExploit
-from machine_learning_ai import MachineLearningAI
-from macos_exploit import MacOSExploit
-from mitm_stingray import MITMStingray
-from network_exploitation import NetworkExploitation
-from predictive_analytics import PredictiveAnalytics
-from proxy_chain_manager import ProxyChainManager
-from real_time_monitoring import RealTimeMonitoring
-from real_time_threat_intelligence import RealTimeThreatIntelligence
-from self_healing_ai_manager import SelfHealingAIManager
-from session_management import SessionManagement
-from settings_manager import SettingsManager
-from threat_intelligence import ThreatIntelligence
-from troubleshooting_manager import TroubleshootingManager
-from vscode_dashboard_manager import VSCodeDashboardManager
-from vulnerability_scanner import VulnerabilityScanner
-from windows_exploit import WindowsExploit
-from wireless_exploitation import WirelessExploitation
-from zero_day_exploits import ZeroDayExploits
-
-class C2Dashboard:
-    # This class integrates with other components like the AI model and chatbot assistant
-    def __init__(self, root):
-        self.root = root
-        self.root.title("C2 Dashboard")
-        self.create_widgets()
-        self.load_user_preferences()
-        self.setup_logging()
-        self.user_role = None
-        self.session_active = False
-        self.chatbot = Chatbot()
-        self.ai_model = AIDeploymentModel("path/to/pretrained/model.h5")
-        self.dark_mode = False
-        self.advanced_decryption = AdvancedDecryption()
-        self.advanced_malware_analysis = AdvancedMalwareAnalysis()
-        self.advanced_social_engineering = AdvancedSocialEngineering()
-        self.adware_manager = AdwareManager()
-        self.ai_red_teaming = AIRedTeaming()
-        self.alerts_notifications = AlertsNotifications()
-        self.android_exploit = AndroidExploit()
-        self.apt_simulation = APTSimulation()
-        self.automated_incident_response = AutomatedIncidentResponse()
-        self.blockchain_logger = BlockchainLogger()
-        self.botnet_manager = BotnetManager()
-        self.data_exfiltration = DataExfiltration()
-        self.data_visualization = DataVisualization()
-        self.deepseek_cody_integration_manager = DeepSeekCodyIntegrationManager()
-        self.device_fingerprinting = DeviceFingerprinting()
-        self.dns_manager = DNSManager()
-        self.download_manager = DownloadManager()
-        self.exploit_payloads = ExploitPayloads()
-        self.fuzzing_engine = FuzzingEngine()
-        self.identity_manager = IdentityManager()
-        self.ios_exploit = IOSExploit()
-        self.iot_exploitation = IoTExploitation()
-        self.linux_exploit = LinuxExploit()
-        self.machine_learning_ai = MachineLearningAI()
-        self.macos_exploit = MacOSExploit()
-        self.mitm_stingray = MITMStingray()
-        self.network_exploitation = NetworkExploitation()
-        self.predictive_analytics = PredictiveAnalytics()
-        self.proxy_chain_manager = ProxyChainManager()
-        self.real_time_monitoring = RealTimeMonitoring()
-        self.real_time_threat_intelligence = RealTimeThreatIntelligence()
-        self.self_healing_ai_manager = SelfHealingAIManager()
-        self.session_management = SessionManagement()
-        self.settings_manager = SettingsManager()
-        self.threat_intelligence = ThreatIntelligence()
-        self.troubleshooting_manager = TroubleshootingManager()
-        self.vscode_dashboard_manager = VSCodeDashboardManager()
-        self.vulnerability_scanner = VulnerabilityScanner()
-        self.windows_exploit = WindowsExploit()
-        self.wireless_exploitation = WirelessExploitation()
-        self.zero_day_exploits = ZeroDayExploits()
-
-    def create_widgets(self):
-        self.tab_control = ttk.Notebook(self.root)
-
-        self.logs_tab = ttk.Frame(self.tab_control)
-        self.exploits_tab = ttk.Frame(self.tab_control)
-        self.communication_tab = ttk.Frame(self.tab_control)
-        self.device_control_tab = ttk.Frame(self.tab_control)
-        self.target_scanning_tab = ttk.Frame(self.tab_control)
-        self.ai_model_tab = ttk.Frame(self.tab_control)
-
-        self.tab_control.add(self.logs_tab, text="Logs")
-        self.tab_control.add(self.exploits_tab, text="Exploits")
-        self.tab_control.add(self.communication_tab, text="Communication")
-        self.tab_control.add(self.device_control_tab, text="Device Control")
-        self.tab_control.add(self.target_scanning_tab, text="Target Scanning")
-        self.tab_control.add(self.ai_model_tab, text="AI Model")
-
-        self.tab_control.pack(expand=1, fill="both")
-
-        self.create_logs_tab()
-        self.create_exploits_tab()
-        self.create_communication_tab()
-        self.create_device_control_tab()
-        self.create_target_scanning_tab()
-        self.create_ai_model_tab()
-
-        self.create_menu()
-        self.add_user_onboarding()
-        self.add_in_app_tutorials()
-
-    def create_menu(self):
-        self.menu_bar = tk.Menu(self.root)
-        self.root.config(menu=self.menu_bar)
-
-        self.file_menu = tk.Menu(self.menu_bar, tearoff=0)
-        self.menu_bar.add_cascade(label="File", menu=self.file_menu)
-        self.file_menu.add_command(label="Exit", command=self.root.quit)
-
-        self.view_menu = tk.Menu(self.menu_bar, tearoff=0)
-        self.menu_bar.add_cascade(label="View", menu=self.view_menu)
-        self.view_menu.add_command(label="Toggle Dark Mode", command=self.toggle_dark_mode)
-
-        self.help_menu = tk.Menu(self.menu_bar, tearoff=0)
-        self.menu_bar.add_cascade(label="Help", menu=self.help_menu)
-        self.help_menu.add_command(label="About", command=self.show_about)
-        self.help_menu.add_command(label="Help", command=self.show_help)
-
-        self.feedback_menu = tk.Menu(self.menu_bar, tearoff=0)
-        self.menu_bar.add_cascade(label="Feedback", menu=self.feedback_menu)
-        self.feedback_menu.add_command(label="Report Issue", command=self.report_issue)
-        self.feedback_menu.add_command(label="Suggest Improvement", command=self.suggest_improvement)
-
-    def toggle_dark_mode(self):
-        self.dark_mode = not self.dark_mode
-        self.apply_theme()
-
-    def apply_theme(self):
-        if self.dark_mode:
-            self.root.tk_setPalette(background='#2e2e2e', foreground='#ffffff', activeBackground='#3e3e3e', activeForeground='#ffffff')
-        else:
-            self.root.tk_setPalette(background='#ffffff', foreground='#000000', activeBackground='#e0e0e0', activeForeground='#000000')
-        self.add_animations_transitions()
-
-    def show_about(self):
-        messagebox.showinfo("About", "C2 Dashboard\nVersion 1.0")
-
-    def show_help(self):
-        messagebox.showinfo("Help", "This is the help section for the C2 Dashboard.")
-
-    def create_logs_tab(self):
-        self.logs_text = tk.Text(self.logs_tab, wrap="word")
-        self.logs_text.pack(expand=1, fill="both")
-
-        self.refresh_logs_button = ttk.Button(self.logs_tab, text="Refresh Logs", command=self.refresh_logs)
-        self.refresh_logs_button.pack()
-
-    def create_exploits_tab(self):
-        self.exploits_listbox = tk.Listbox(self.exploits_tab)
-        self.exploits_listbox.pack(expand=1, fill="both")
-
-        self.load_exploits_button = ttk.Button(self.exploits_tab, text="Load Exploits", command=self.load_exploits)
-        self.load_exploits_button.pack()
-
-        self.run_exploit_button = ttk.Button(self.exploits_tab, text="Run Exploit", command=self.run_exploit)
-        self.run_exploit_button.pack()
-
-    def create_communication_tab(self):
-        self.communication_text = tk.Text(self.communication_tab, wrap="word")
-        self.communication_text.pack(expand=1, fill="both")
-
-        self.send_message_button = ttk.Button(self.communication_tab, text="Send Message", command=self.send_message)
-        self.send_message_button.pack()
-
-    def create_device_control_tab(self):
-        self.device_control_text = tk.Text(self.device_control_tab, wrap="word")
-        self.device_control_text.pack(expand=1, fill="both")
-
-        self.deploy_exploit_button = ttk.Button(self.device_control_tab, text="Deploy Exploit", command=self.deploy_exploit)
-        self.deploy_exploit_button.pack()
-
-    def create_target_scanning_tab(self):
-        self.target_scanning_text = tk.Text(self.target_scanning_tab, wrap="word")
-        self.target_scanning_text.pack(expand=1, fill="both")
-
-        self.scan_targets_button = ttk.Button(self.target_scanning_tab, text="Scan Targets", command=self.scan_targets)
-        self.scan_targets_button.pack()
-
-    def create_ai_model_tab(self):
-        self.ai_model_input_text = tk.Text(self.ai_model_tab, wrap="word")
-        self.ai_model_input_text.pack(expand=1, fill="both")
-
-        self.predict_button = ttk.Button(self.ai_model_tab, text="Predict", command=self.predict)
-        self.predict_button.pack()
-
-        self.ai_model_output_text = tk.Text(self.ai_model_tab, wrap="word")
-        self.ai_model_output_text.pack(expand=1, fill="both")
-
-    def refresh_logs(self):
-        self.logs_text.delete(1.0, tk.END)
-        with open("logs/deployment.log", "r") as f:
-            logs = f.read()
-            self.logs_text.insert(tk.END, logs)
-
-    def load_exploits(self):
-        self.exploits_listbox.delete(0, tk.END)
-        exploits = os.listdir("exploits")
-        for exploit in exploits:
-            self.exploits_listbox.insert(tk.END, exploit)
-
-    def run_exploit(self):
-        selected_exploit = self.exploits_listbox.get(tk.ACTIVE)
-        if selected_exploit:
-            exploit_path = os.path.join("exploits", selected_exploit)
-            result = subprocess.run([exploit_path], capture_output=True, text=True)
-            messagebox.showinfo("Exploit Result", result.stdout)
-
-    def send_message(self):
-        message = self.communication_text.get(1.0, tk.END).strip()
-        if message:
-            # Implement secure communication logic here
-            messagebox.showinfo("Message Sent", "Message sent successfully!")
-
-    def deploy_exploit(self):
-        device_info = self.device_control_text.get(1.0, tk.END).strip()
-        if device_info:
-            phone_regex = re.compile(r'\+?1?\d{9,15}')
-            email_regex = re.compile(r'[\w\.-]+@[\w\.-]+')
-            ip_port_regex = re.compile(r'(\d{1,3}\.){3}\d{1,3}:\d{1,3}')
-
-            phone_numbers = phone_regex.findall(device_info)
-            emails = email_regex.findall(device_info)
-            ip_ports = ip_port_regex.findall(device_info)
-
-            # Implement exploit deployment logic based on extracted information
-            messagebox.showinfo("Exploit Deployment", "Exploits deployed successfully!")
-
-    def scan_targets(self):
-        shodan_api_key = os.getenv("SHODAN_API_KEY")
-        nmap_api_key = os.getenv("NMAP_API_KEY")
-        shodan_api = shodan.Shodan(shodan_api_key)
-        nm = nmap.PortScanner()
-
-        try:
-            results = shodan_api.search('default password')
-            for result in results['matches']:
-                ip = result['ip_str']
-                nm.scan(ip, '22-443')
-                self.target_scanning_text.insert(tk.END, f"IP: {ip}\n")
-                for proto in nm[ip].all_protocols():
-                    lport = nm[ip][proto].keys()
-                    for port in lport:
-                        self.target_scanning_text.insert(tk.END, f"Port: {port}\tState: {nm[ip][proto][port]['state']}\n")
-        except shodan.APIError as e:
-            messagebox.showerror("Shodan Error", str(e))
-
-    def predict(self):
-        input_data = self.ai_model_input_text.get(1.0, tk.END).strip().split('\n')
-        predictions = self.ai_model.predict(input_data)
-        self.ai_model_output_text.delete(1.0, tk.END)
-        self.ai_model_output_text.insert(tk.END, str(predictions))
-
-    def setup_logging(self):
-        logging.basicConfig(filename='logs/gui.log', level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-
-    def load_user_preferences(self):
-        try:
-            with open('config.json', 'r') as f:
-                self.user_preferences = json.load(f)
-        except FileNotFoundError:
-            self.user_preferences = {}
-
-    def save_user_preferences(self):
-        with open('config.json', 'w') as f:
-            json.dump(self.user_preferences, f)
-
-    def login(self):
-        username = askstring("Login", "Enter your username:")
-        password = askstring("Login", "Enter your password:", show='*')
-        if self.authenticate_user(username, password):
-            self.user_role = self.get_user_role(username)
-            self.session_active = True
-            self.root.after(60000, self.check_session_timeout)
-            self.implement_2fa()
-        else:
-            messagebox.showerror("Login Failed", "Invalid credentials")
-
-    def authenticate_user(self, username, password):
-        # Implement user authentication logic here
-        return True
-
-    def get_user_role(self, username):
-        # Implement user role retrieval logic here
-        return "admin"
-
-    def check_session_timeout(self):
-        if self.session_active:
-            self.session_active = False
-            messagebox.showinfo("Session Timeout", "You have been logged out due to inactivity")
-            self.login()
-
-    def run_post_exploitation_module(self, module_name):
-        # Implement post-exploitation module execution logic here
-        messagebox.showinfo("Post-Exploitation Module", f"{module_name} executed successfully")
-
-    def setup_ddns(self):
-        no_ip_username = os.getenv("NO_IP_USERNAME")
-        no_ip_password = os.getenv("NO_IP_PASSWORD")
-        no_ip_hostname = os.getenv("NO_IP_HOSTNAME")
-
-        if not no_ip_username or not no_ip_password or not no_ip_hostname:
-            messagebox.showerror("DDNS Error", "No-IP DDNS credentials are missing")
-            return
-
-        update_url = f"https://{no_ip_username}:{no_ip_password}@dynupdate.no-ip.com/nic/update?hostname={no_ip_hostname}"
-        response = requests.get(update_url)
-
-        if response.status_code == 200:
-            messagebox.showinfo("DDNS Update", "No-IP DDNS update successful")
-        else:
-            messagebox.showerror("DDNS Update", f"No-IP DDNS update failed: {response.text}")
-
-    def setup_reverse_dns_tunneling(self):
-        # Implement reverse DNS tunneling setup logic here
-        pass
-
-    def integrate_chatbot(self):
-        self.chatbot_popup = tk.Toplevel(self.root)
-        self.chatbot_popup.title("Chatbot")
-        self.chatbot_text = tk.Text(self.chatbot_popup, wrap="word")
-        self.chatbot_text.pack(expand=1, fill="both")
-        self.chatbot_entry = tk.Entry(self.chatbot_popup)
-        self.chatbot_entry.pack(fill="x")
-        self.chatbot_entry.bind("<Return>", self.send_chatbot_command)
-
-    def send_chatbot_command(self, event):
-        command = self.chatbot_entry.get()
-        if command:
-            response = self.chatbot.process_command(command)
-            self.chatbot_text.insert(tk.END, f"User: {command}\n")
-            self.chatbot_text.insert(tk.END, f"Chatbot: {response}\n")
-            self.chatbot_entry.delete(0, tk.END)
-
-    def spoof_sms(self, phone_number, message):
-        # Implement SMS spoofing logic here
-        pass
-
-    def spoof_email(self, email_address, subject, message):
-        # Implement email spoofing logic here
-        pass
-
-    def prompt_ai_scan_targets(self):
-        self.chatbot_text.insert(tk.END, "Prompting AI to scan targets...\n")
-        self.ai_model.scan_targets()
-        self.chatbot_text.insert(tk.END, "AI scan targets completed.\n")
-
-    def prompt_ai_modify_exploits(self, target_info):
-        self.chatbot_text.insert(tk.END, "Prompting AI to modify exploits...\n")
-        self.ai_model.modify_exploits(target_info)
-        self.chatbot_text.insert(tk.END, "AI modify exploits completed.\n")
-
-    def prompt_ai_deploy_exploits(self, target_info):
-        self.chatbot_text.insert(tk.END, "Prompting AI to deploy exploits...\n")
-        self.ai_model.deploy_exploit(target_info)
-        self.chatbot_text.insert(tk.END, "AI deploy exploits completed.\n")
-
-    def prompt_ai_post_exploitation(self, module_name):
-        self.chatbot_text.insert(tk.END, "Prompting AI to run post-exploitation module...\n")
-        self.run_post_exploitation_module(module_name)
-        self.chatbot_text.insert(tk.END, "AI post-exploitation module completed.\n")
-
-    def add_tooltips(self):
-        # Add tooltips to various widgets
-        pass
-
-    def add_help_sections(self):
-        # Add help sections to guide users through the app's features
-        pass
-
-    def add_user_onboarding(self):
-        # Add a user onboarding process
-        pass
-
-    def add_in_app_tutorials(self):
-        # Implement in-app tutorials and guides
-        pass
-
-    def add_feedback_system(self):
-        # Add a feedback system for users to report issues and suggest improvements
-        pass
-
-    def add_animations_transitions(self):
-        # Add animations and transitions for a smooth user experience
-        pass
-
-    def implement_2fa(self):
-        # Implement two-factor authentication (2FA) for user login
-        pass
-
-    def add_encryption(self):
-        # Add encryption for sensitive data stored in the app
-        pass
-
-    def integrate_secure_communication(self):
-        # Integrate a secure communication protocol for data transmission
-        pass
-
-    def implement_session_timeout(self):
-        # Implement a session timeout feature to automatically log out inactive users
-        pass
-
-    def add_support_for_more_exploit_types(self):
-        # Add support for more exploit types and platforms
-        pass
-
-    def integrate_vulnerability_scanner(self):
-        # Integrate a vulnerability scanner to identify potential security issues in target systems
-        pass
-
-    def implement_reporting_feature(self):
-        # Implement a reporting feature to generate detailed reports on exploit activities and results
-        pass
-
-    def add_notification_system(self):
-        # Add a notification system to alert users of important events or updates within the app
-        pass
-
-    def integrate_chatbot_assistant(self):
-        # Integrate a chatbot to assist users with common tasks and provide guidance
-        pass
-
-    def add_multimedia_support(self):
-        # Add support for multimedia messages, such as images, videos, and files
-        pass
-
-    def implement_message_encryption(self):
-        # Implement message encryption to ensure secure communication
-        pass
-
-    def add_search_feature(self):
-        # Add a search feature to quickly find specific messages or conversations
-        pass
-
-    def enable_message_reactions(self):
-        # Enable message reactions and emojis for better user interaction
-        pass
-
-if __name__ == "__main__":
-    root = tk.Tk()
-    app = C2Dashboard(root)
-    app.login()
-    app.setup_ddns()
-    app.setup_reverse_dns_tunneling()
-    app.integrate_chatbot()
-    root.mainloop()
